refactor: use the new com.hlag.tools.commvis:api library
