+++ analyzer/pom.xml
@@ -42,6 +42,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.hlag.tools.commvis</groupId>
+            <artifactId>api</artifactId>
+            <version>1.1.0</version>
+        </dependency>
+        <dependency>
             <groupId>javax.ejb</groupId>
             <artifactId>javax.ejb-api</artifactId>
             <version>3.2.2</version>
