+++ analyzer/src/main/java/com/hlag/tools/commvis/domain/command/ScannerCommand.java
@@ -1,8 +1,8 @@
 package com.hlag.tools.commvis.domain.command;
 
-import com.hlag.tools.commvis.domain.model.CommunicationModel;
-import com.hlag.tools.commvis.domain.model.IEndpoint;
-import com.hlag.tools.commvis.service.IEndpointScannerService;
+import com.hlag.tools.commvis.analyzer.model.CommunicationModel;
+import com.hlag.tools.commvis.analyzer.model.ISenderReceiverCommunication;
+import com.hlag.tools.commvis.analyzer.service.IScannerService;
 import com.hlag.tools.commvis.service.IExportModelService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -17,7 +17,7 @@ import java.util.concurrent.Callable;
 @Component
 @CommandLine.Command(name = "DistCommVis", mixinStandardHelpOptions = true, description = "Analyzes the classpath and extracts endpoints and event sender/receiver.")
 public class ScannerCommand implements Callable<Integer> {
-    private final IEndpointScannerService[] scannerServices;
+    private final IScannerService[] scannerServices;
     private final IExportModelService[] exportModelServices;
 
     @CommandLine.Parameters(index = "1", description = "The root package to analyze (including all sub packages)")
@@ -29,16 +29,15 @@ public class ScannerCommand implements Callable<Integer> {
     @CommandLine.Option(names = {"-n", "--name"}, defaultValue = "application", description = "Name of the project to be analyzed.")
     private String projectName;
 
-    public ScannerCommand(IEndpointScannerService[] scannerServices, IExportModelService[] exportModelServices) {
+    public ScannerCommand(IScannerService[] scannerServices, IExportModelService[] exportModelServices) {
         this.scannerServices = scannerServices;
         this.exportModelServices = exportModelServices;
     }
 
     @Override
     public Integer call() {
-        Collection<IEndpoint> endpoints = new HashSet<>();
-        Arrays.asList(scannerServices).forEach(s -> endpoints.addAll(s.scanClasspath(rootPackageName)));
-
+        Collection<ISenderReceiverCommunication> endpoints = new HashSet<>();
+        Arrays.asList(scannerServices).forEach(s -> endpoints.addAll(s.scanSenderAndReceiver(rootPackageName)));
         CommunicationModel model = new CommunicationModel(projectId, projectName, endpoints);
         Arrays.asList(exportModelServices).forEach(s -> s.export(model, "model"));
