--- analyzer/src/main/java/com/hlag/tools/commvis/domain/model/AbstractCommunicationModelVisitor.java
@@ -1,11 +1,0 @@
-package com.hlag.tools.commvis.domain.model;
-
-/**
- * The visitor pattern to transform the internal model.
- */
-public abstract class AbstractCommunicationModelVisitor {
-    public abstract void visit(CommunicationModel model);
-
-    public abstract void visit(HttpEndpoint endpoint);
-    public abstract void visit(JmsEndpoint endpoint);
-}
