--- analyzer/src/main/java/com/hlag/tools/commvis/domain/model/HttpEndpoint.java
@@ -1,23 +1,0 @@
-package com.hlag.tools.commvis.domain.model;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@AllArgsConstructor
-@Getter
-@ToString
-@EqualsAndHashCode
-public class HttpEndpoint implements IEndpoint {
-    private String className;
-    private String methodName;
-
-    private String type;
-    private String path;
-
-    @Override
-    public void visit(AbstractCommunicationModelVisitor visitor) {
-        visitor.visit(this);
-    }
-}
