--- analyzer/src/main/java/com/hlag/tools/commvis/domain/model/IEndpoint.java
@@ -1,7 +1,0 @@
-package com.hlag.tools.commvis.domain.model;
-
-public interface IEndpoint {
-    String getClassName();
-
-    void visit(AbstractCommunicationModelVisitor visitor);
-}
