+++ analyzer/src/test/java/com/hlag/tools/commvis/domain/command/ScannerCommandTest.java
@@ -1,6 +1,6 @@
 package com.hlag.tools.commvis.domain.command;
 
-import com.hlag.tools.commvis.service.IEndpointScannerService;
+import com.hlag.tools.commvis.analyzer.service.IScannerService;
 import com.hlag.tools.commvis.service.IExportModelService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -14,7 +14,7 @@ class ScannerCommandTest {
     @Mock
     IExportModelService exportModelServices;
     @Mock
-    private IEndpointScannerService scannerService;
+    private IScannerService scannerService;
 
     @BeforeEach
     public void init() {
@@ -23,21 +23,21 @@ class ScannerCommandTest {
 
     @Test
     void shouldAcceptProjectIdAndPackageName_whenCalledFromCommandLine() {
-        int actualExitCode = new CommandLine(new ScannerCommand(new IEndpointScannerService[]{scannerService}, new IExportModelService[]{exportModelServices})).execute("4711", "com.hlag");
+        int actualExitCode = new CommandLine(new ScannerCommand(new IScannerService[]{scannerService}, new IExportModelService[]{exportModelServices})).execute("4711", "com.hlag");
 
         assertThat(actualExitCode).isZero();
     }
 
     @Test
     void shouldExitWithStatusCode2_whenCalledFromCommandLine_givenNoParameters() {
-        int actualExitCode = new CommandLine(new ScannerCommand(new IEndpointScannerService[]{scannerService}, new IExportModelService[]{exportModelServices})).execute();
+        int actualExitCode = new CommandLine(new ScannerCommand(new IScannerService[]{scannerService}, new IExportModelService[]{exportModelServices})).execute();
 
         assertThat(actualExitCode).isEqualTo(2);
     }
 
     @Test
     void shouldExitWithStatusCode2_whenCalledFromCommandLine_givenProjectIdOnly() {
-        int actualExitCode = new CommandLine(new ScannerCommand(new IEndpointScannerService[]{scannerService}, new IExportModelService[]{exportModelServices})).execute("4711");
+        int actualExitCode = new CommandLine(new ScannerCommand(new IScannerService[]{scannerService}, new IExportModelService[]{exportModelServices})).execute("4711");
 
         assertThat(actualExitCode).isEqualTo(2);
     }
